server/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
awsebcli==3.20.5
bcrypt==4.0.1
blessed==1.20.0
botocore==1.29.81
cached-property==1.5.2
cement==2.8.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.3
cryptography==40.0.2
dataclasses-json==0.5.7
dnspython==2.3.0
docker==4.4.4
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
Flask==2.0.3
Flask-Cors==3.0.10
frozenlist==1.3.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==3.2.0
langchain==0.0.147
loguru==0.7.0
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.24.3
openai==0.27.4
openapi-schema-pydantic==1.2.4
packaging==23.1
paramiko==3.1.0
pathspec==0.10.1
pinecone-client==2.2.1
pycparser==2.21
pydantic==1.10.7
PyNaCl==1.5.0
pyrsistent==0.19.3
python-dateutil==2.8.2
PyYAML==5.4.1
regex==2023.3.23
requests==2.26.0
semantic-version==2.8.5
six==1.14.0
SQLAlchemy==1.4.47
tenacity==8.2.2
termcolor==1.1.0
texttable==1.6.7
tiktoken==0.3.3
tqdm==4.65.0
typing-inspect==0.8.0
typing_extensions==4.5.0
urllib3==1.26.15
wcwidth==0.1.9
websocket-client==0.59.0
Werkzeug==2.2.3
yarl==1.9.1
youtube-transcript-api==0.6.0
